Single ASAM callout bakground color to black
The single ASAM callout had a black view but the callout background was white.  This meant the triangle/arrow pointing to the annotation was white.  Changed to black to match view.

Also fixed the callout deselect.  After selecting a callout to viwe ASAMs the callout was never deselected and would apper grey